% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcular_extensao}
\alias{calcular_extensao}
\title{Calcula a extensão em km² para uma zona UTM.}
\usage{
calcular_extensao(shapefile, epsg)
}
\arguments{
\item{shapefile}{Um objeto do tipo sf (simple features) representando o shapefile.}

\item{epsg}{Um caractere representando o código EPSG da zona UTM.}
}
\value{
Um número representando a área em km² do shapefile na zona UTM especificada.
}
\description{
Esta função transforma um shapefile para uma zona UTM específica e calcula
a área do shapefile na nova projeção.
}
