% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{calcular_numero_tiles}
\alias{calcular_numero_tiles}
\title{Calcular Número de Tiles}
\usage{
calcular_numero_tiles(
  caminho_raster,
  caminho_limite,
  numcol,
  numrow,
  overwrite
)
}
\arguments{
\item{caminho_raster}{O caminho para o arquivo raster (ex: .tif).}

\item{caminho_limite}{O caminho para o shapefile que define o limite da área.}

\item{numcol}{O número de colunas de cada tile.}

\item{numrow}{O número de linhas de cada tile.}
}
\value{
O número total de tiles que podem ser criados dentro do limite do polígono.
}
\description{
Esta função calcula o número de tiles que podem ser criados com base no limite do polígono
e nas dimensões de cada tile (número de colunas e linhas).
}
\examples{
\dontrun{
num_tiles <- calcular_numero_tiles('./copernicusDem/copernicusDem_Mg.tif',
                                    './limite_area/mg_buf.shp',
                                    numcol = 3600, numrow = 3600)
print(num_tiles)}
}
