% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_convergence_index}
\alias{calculate_convergence_index}
\title{Calculate Convergence Index}
\usage{
calculate_convergence_index(
  dem,
  outdir,
  prefix,
  method = "Aspect",
  neighbours = 1,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the convergence index file.}

\item{method}{The method to use for calculation. Default is "Aspect".}

\item{neighbours}{The number of neighbours to consider. Default is 1.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating the convergence index.
}
\description{
Computes the convergence index from a DEM using SAGA GIS.
}
