% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_curvature_classification}
\alias{calculate_curvature_classification}
\title{Calculate Curvature Classification}
\usage{
calculate_curvature_classification(
  dem,
  outdir,
  prefix,
  threshold = 0.05,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the curvature classification file.}

\item{threshold}{The threshold for curvature classification. Default is 0.05.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating curvature classification.
}
\description{
Computes curvature classification from a DEM using SAGA GIS.
}
