% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_diurnal_anisotropic_heat}
\alias{calculate_diurnal_anisotropic_heat}
\title{Calculate Diurnal Anisotropic Heat}
\usage{
calculate_diurnal_anisotropic_heat(
  dem,
  outdir,
  alpha_max = 202.5,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the diurnal anisotropic heat file.}

\item{alpha_max}{The maximum alpha value for the calculation. Default is 202.5.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating diurnal anisotropic heat data.
}
\description{
Computes diurnal anisotropic heat from a DEM using SAGA GIS.
}
