% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_downslope_distance_gradient}
\alias{calculate_downslope_distance_gradient}
\title{Calculate Downslope Distance Gradient}
\usage{
calculate_downslope_distance_gradient(
  dem,
  outdir,
  distance = 10,
  output = 2,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the downslope distance gradient files.}

\item{distance}{The distance parameter for the calculation. Default is 10.}

\item{output}{The output type for the calculation. Default is 2.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating downslope distance gradient data.
}
\description{
Computes downslope distance gradient from a DEM using SAGA GIS.
}
