% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_effective_air_flow_heights}
\alias{calculate_effective_air_flow_heights}
\title{Calculate Effective Air Flow Heights}
\usage{
calculate_effective_air_flow_heights(
  dem,
  outdir,
  prefix,
  num_tile,
  luv = 1,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the effective air flow heights file.}

\item{luv}{The LUV parameter for the calculation. Default is 1.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating effective air flow heights data.
}
\description{
Computes effective air flow heights from a DEM using SAGA GIS.
}
