% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_hillshade}
\alias{calculate_hillshade}
\title{Calculate Hillshade}
\usage{
calculate_hillshade(
  dem,
  outdir,
  prefix,
  azimuth = 315,
  declination = 45,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the hillshade file.}

\item{azimuth}{The azimuth angle for the light source. Default is 315.}

\item{declination}{The declination angle for the light source. Default is 45.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating hillshade.
}
\description{
Computes hillshade from a Digital Elevation Model (DEM) using SAGA GIS.
}
