% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_mass_balance_index}
\alias{calculate_mass_balance_index}
\title{Calculate Mass Balance Index}
\usage{
calculate_mass_balance_index(
  dem,
  outdir,
  tslope = 15,
  tcurve = 0.01,
  threl = 15,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the mass balance index file.}

\item{tslope}{The slope threshold for the calculation. Default is 15.}

\item{tcurve}{The curvature threshold for the calculation. Default is 0.01.}

\item{threl}{The relative threshold for the calculation. Default is 15.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating mass balance index data.
}
\description{
Computes mass balance index from a DEM using SAGA GIS.
}
