% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_morphometric_protection_index}
\alias{calculate_morphometric_protection_index}
\title{Calculate Morphometric Protection Index}
\usage{
calculate_morphometric_protection_index(
  dem,
  outdir,
  num_tile,
  prefix,
  radius = 2000,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the morphometric protection index file.}

\item{radius}{The radius parameter for the calculation. Default is 2000.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating morphometric protection index data.
}
\description{
Computes morphometric protection index from a DEM using SAGA GIS.
}
