% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_real_surface_area}
\alias{calculate_real_surface_area}
\title{Calculate Real Surface Area}
\usage{
calculate_real_surface_area(dem, outdir, verbose = TRUE, env)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the real surface area file.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating real surface area data.
}
\description{
Computes real surface area from a DEM using SAGA GIS.
}
