% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_relative_heights_and_slope_positions}
\alias{calculate_relative_heights_and_slope_positions}
\title{Calculate Relative Heights and Slope Positions}
\usage{
calculate_relative_heights_and_slope_positions(
  dem,
  outdir,
  w = 0.5,
  t = 10,
  e = 2,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the relative heights and slope positions file.}

\item{w}{The weight parameter for the calculation. Default is 0.5.}

\item{t}{The threshold parameter for the calculation. Default is 10.}

\item{e}{The exponent parameter for the calculation. Default is 2.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating relative heights and slope positions data.
}
\description{
Computes relative heights and slope positions from a DEM using SAGA GIS.
}
