% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_ridge_level}
\alias{calculate_ridge_level}
\title{Calculate Ridge Level}
\usage{
calculate_ridge_level(
  dem,
  outdir,
  prefix,
  num_tile,
  threshold = 1,
  nounderground = 1,
  order = 4,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the ridge level file.}

\item{threshold}{The threshold parameter for the calculation. Default is 1.}

\item{nounderground}{The no underground parameter for the calculation. Default is 1.}

\item{order}{The order parameter for the calculation. Default is 4.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating ridge level data.
}
\description{
Computes ridge level from a DEM using SAGA GIS.
}
