% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_sky_view_factor}
\alias{calculate_sky_view_factor}
\title{Calculate Sky View Factor}
\usage{
calculate_sky_view_factor(
  dem,
  outdir,
  prefix,
  num_tile,
  radius = 10000,
  method = 1,
  dlevel = 3,
  ndirs = 8,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the sky view factor file.}

\item{radius}{The radius parameter for the calculation. Default is 10000.}

\item{method}{The method parameter for the calculation. Default is 1.}

\item{dlevel}{The detail level parameter for the calculation. Default is 3.}

\item{ndirs}{The number of directions parameter for the calculation. Default is 8.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating sky view factor data.
}
\description{
Computes sky view factor from a DEM using SAGA GIS.
}
