% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_slope_aspect_curvatures}
\alias{calculate_slope_aspect_curvatures}
\title{Calculate Slope, Aspect, and Curvatures}
\usage{
calculate_slope_aspect_curvatures(
  dem,
  outdir,
  prefix,
  num_tile,
  method = 4,
  unit_slope = 1,
  unit_aspect = 1,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the slope, aspect, and curvatures files.}

\item{method}{The method parameter for the calculation. Default is 4.}

\item{unit_slope}{The unit slope parameter for the calculation. Default is 1.}

\item{unit_aspect}{The unit aspect parameter for the calculation. Default is 1.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating slope, aspect, and curvatures data.
}
\description{
Computes slope, aspect, and curvatures from a DEM using SAGA GIS.
}
