% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_solar_radiation}
\alias{calculate_solar_radiation}
\title{Calculate Solar Radiation}
\usage{
calculate_solar_radiation(
  dem,
  outdir,
  prefix,
  latitude,
  start_date = "01-01-2018",
  end_date = "12-31-2018",
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the solar radiation file.}

\item{latitude}{The latitude for the location.}

\item{start_date}{The start date for the calculation period. Default is "01-01-2018".}

\item{end_date}{The end date for the calculation period. Default is "12-31-2018".}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating solar radiation data.
}
\description{
Computes potential incoming solar radiation from a DEM using SAGA GIS.
}
