% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_surface_specific_points}
\alias{calculate_surface_specific_points}
\title{Calculate Surface Specific Points}
\usage{
calculate_surface_specific_points(
  dem,
  outdir,
  prefix,
  num_tile,
  method = 1,
  threshold = 2,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the surface specific points file.}

\item{method}{The method parameter for the calculation. Default is 1.}

\item{threshold}{The threshold parameter for the calculation. Default is 2.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating surface specific points data.
}
\description{
Computes surface specific points from a DEM using SAGA GIS.
}
