% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_terrain_surface_classification}
\alias{calculate_terrain_surface_classification}
\title{Calculate Terrain Surface Classification}
\usage{
calculate_terrain_surface_classification(
  dem,
  outdir,
  prefix,
  num_tile,
  type = 2,
  conv_scale = 10,
  text_scale = 2,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the terrain surface classification file.}

\item{type}{The type parameter for the calculation. Default is 2.}

\item{conv_scale}{The convolution scale parameter for the calculation. Default is 10.}

\item{text_scale}{The texture scale parameter for the calculation. Default is 2.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating terrain surface classification data.
}
\description{
Computes terrain surface classification from a DEM using SAGA GIS.
}
