% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_topographic_position_index}
\alias{calculate_topographic_position_index}
\title{Calculate Topographic Position Index}
\usage{
calculate_topographic_position_index(
  dem,
  outdir,
  prefix,
  num_tile,
  radius_min = 0,
  radius_max = 100,
  dw_weighting = 0,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the topographic position index file.}

\item{radius_min}{The minimum radius parameter for the calculation. Default is 0.}

\item{radius_max}{The maximum radius parameter for the calculation. Default is 100.}

\item{dw_weighting}{The distance weighting parameter for the calculation. Default is 0.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating topographic position index data.
}
\description{
Computes topographic position index from a DEM using SAGA GIS.
}
