% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_tpi_based_landform_classification}
\alias{calculate_tpi_based_landform_classification}
\title{Calculate TPI Based Landform Classification}
\usage{
calculate_tpi_based_landform_classification(
  dem,
  outdir,
  prefix,
  num_tile,
  radius_a_min = 0,
  radius_a_max = 100,
  radius_b_min = 0,
  radius_b_max = 1000,
  dw_weighting = 0,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the TPI based landform classification file.}

\item{radius_a_min}{The minimum radius A parameter for the calculation. Default is 0.}

\item{radius_a_max}{The maximum radius A parameter for the calculation. Default is 100.}

\item{radius_b_min}{The minimum radius B parameter for the calculation. Default is 0.}

\item{radius_b_max}{The maximum radius B parameter for the calculation. Default is 1000.}

\item{dw_weighting}{The distance weighting parameter for the calculation. Default is 0.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating TPI based landform classification data.
}
\description{
Computes TPI based landform classification from a DEM using SAGA GIS.
}
