% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_vector_ruggedness_measure}
\alias{calculate_vector_ruggedness_measure}
\title{Calculate Vector Ruggedness Measure}
\usage{
calculate_vector_ruggedness_measure(
  dem,
  outdir,
  prefix,
  num_tile,
  mode = 1,
  radius = 1,
  dw_weighting = 0,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the vector ruggedness measure file.}

\item{mode}{The mode parameter for the calculation. Default is 1.}

\item{radius}{The radius parameter for the calculation. Default is 1.}

\item{dw_weighting}{The distance weighting parameter for the calculation. Default is 0.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating vector ruggedness measure data.
}
\description{
Computes vector ruggedness measure from a DEM using SAGA GIS.
}
