% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_wetness_index}
\alias{calculate_wetness_index}
\title{Calculate Wetness Index}
\usage{
calculate_wetness_index(
  dem,
  outdir,
  prefix,
  num_tile,
  suction = 10,
  area_type = 1,
  slope_type = 1,
  slope_min = 0,
  slope_off = 0.1,
  slope_weight = 1,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the wetness index file.}

\item{suction}{The suction parameter for the calculation. Default is 10.}

\item{area_type}{The area type parameter for the calculation. Default is 1.}

\item{slope_type}{The slope type parameter for the calculation. Default is 1.}

\item{slope_min}{The minimum slope parameter for the calculation. Default is 0.}

\item{slope_off}{The slope offset parameter for the calculation. Default is 0.1.}

\item{slope_weight}{The slope weight parameter for the calculation. Default is 1.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating wetness index data.
}
\description{
Computes wetness index from a DEM using SAGA GIS.
}
