% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{calculate_wind_exposition_index}
\alias{calculate_wind_exposition_index}
\title{Calculate Wind Exposition Index}
\usage{
calculate_wind_exposition_index(
  dem,
  outdir,
  prefix,
  num_tile,
  maxdist = 300,
  step = 15,
  oldver = 0,
  accel = 1.5,
  pyramids = 0,
  verbose = TRUE,
  env
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the wind exposition index file.}

\item{maxdist}{The maximum distance parameter for the calculation. Default is 300.}

\item{step}{The step parameter for the calculation. Default is 15.}

\item{oldver}{The old version parameter for the calculation. Default is 0.}

\item{accel}{The acceleration parameter for the calculation. Default is 1.5.}

\item{pyramids}{The pyramids parameter for the calculation. Default is 0.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{env}{The SAGA environment settings.}
}
\value{
None. The function is used for its side effects of generating wind exposition index data.
}
\description{
Computes wind exposition index from a DEM using SAGA GIS.
}
