% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{criar_grid_tiles}
\alias{criar_grid_tiles}
\title{Criar grid de Tiles}
\usage{
criar_grid_tiles(
  path_raster,
  shape_limite,
  numcol = 3600,
  numrow = 3600,
  path_saida
)
}
\arguments{
\item{numcol}{numero de linhas do tile}

\item{numrow}{número de colunas do tile}

\item{caminho_raster}{O caminho para o arquivo raster (ex: .tif).Raster deve estar
projeção geografica com datum SIRGAS2000 ou WGS84}

\item{caminho_limite}{O caminho para o shapefile que define o limite da área.}

\item{caminho_saida}{O caminho onde o shapefile dos tiles cortados será salvo.}
}
\value{
Um objeto sf contendo os tiles cortados.
}
\description{
Esta função cria uma grade de tiles a partir de um raster e corta os tiles
pelo limite de uma área definida em um shapefile.
}
\examples{
\dontrun {
criar_grid_tiles('./copernicusDem/copernicusDem_Mg.tif',
                 './limite_area/mg_buf.shp',
                 './saida/tiles_cortados.shp')}


}
