% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{determinar_epsg_utm}
\alias{determinar_epsg_utm}
\title{Determina o EPSG UTM a partir das coordenadas geográficas de um shapefile.}
\usage{
determinar_epsg_utm(shapefile)
}
\arguments{
\item{shapefile}{Um objeto do tipo sf (simple features) representando o shapefile.}
}
\value{
Um caractere representando o código EPSG da zona UTM com a maior área.
}
\description{
Esta função calcula a zona UTM correspondente às coordenadas geográficas
de um shapefile e retorna o código EPSG da zona que contém a maior área
do shapefile, considerando as zonas adjacentes se necessário.
}
\examples{

# Supondo que 'meu_shapefile' é um objeto sf válido
epsg <- determinar_epsg_utm(meu_shapefile)
print(epsg)

}
