% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{install_and_load}
\alias{install_and_load}
\title{This function checks if each package in the provided list is installed. If a package is not installed, it installs the package and its dependencies. Then, it loads the package into the R session.}
\usage{
install_and_load(packages)
}
\arguments{
\item{packages}{A character vector of package names to be installed and loaded.}
}
\value{
None. The function is used for its side effects of installing and loading packages.
}
\description{
This function checks if each package in the provided list is installed. If a package is not installed, it installs the package and its dependencies. Then, it loads the package into the R session.
}
