% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{process_dem}
\alias{process_dem}
\title{Main Processing Function with Parallel Execution}
\usage{
process_dem(
  dem,
  outdir,
  prefix = "morpho",
  cores = 2,
  parallel = TRUE,
  verbose = TRUE
)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the processed files.}

\item{cores}{The number of cores to use for parallel processing. Default is 2.}

\item{parallel}{Logical indicating whether to enable parallel processing. Default is TRUE.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}
}
\value{
None. The function is used for its side effects of processing the DEM.
}
\description{
Processes a DEM using multiple terrain analysis functions in parallel.
}
