% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{process_dem_area}
\alias{process_dem_area}
\title{Processamento de DEMs por area}
\usage{
process_dem_area(
  dem_files,
  outdir,
  prefixo,
  cores = 1,
  parallel = FALSE,
  verbose = TRUE
)
}
\arguments{
\item{dem_files}{Um vetor de strings contendo os caminhos dos arquivos raster (DEMs) a serem processados.}

\item{outdir}{O diretório onde os resultados processados serão salvos.}

\item{prefixo}{Um prefixo a ser adicionado aos nomes dos arquivos de saída.}

\item{cores}{Um número inteiro especificando o número de núcleos a serem usados para processamento paralelo. O padrão é 1.}

\item{parallel}{Um valor lógico que indica se o processamento deve ser feito em paralelo. O padrão é FALSE.}

\item{verbose}{Um valor lógico que indica se mensagens de progresso devem ser exibidas. O padrão é TRUE.}
}
\value{
Esta função não retorna um valor, mas salva os resultados processados no diretório especificado.
}
\description{
Esta função aplica uma série de funções de processamento a arquivos raster (DEMs) em um diretório especificado.
Cada função é aplicada a todos os rasters antes de passar para a próxima função.
}
\examples{
# Definir o diretório onde os rasters estão localizados
raster_dir <- "caminho/para/sua/pasta"
dem_files <- list.files(raster_dir, pattern = "\\\\.tif$", full.names = TRUE)

# Chamar a função process_dem
process_dem_area(dem_files, outdir = "caminho/para/diretorio/de/saida", prefixo = "resultado", cores = 4, parallel = TRUE)

}
