% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{process_dem_area2}
\alias{process_dem_area2}
\title{Processamento de Área DEM}
\usage{
process_dem_area2(
  dem_files,
  outdir,
  prefixo,
  functions_to_run,
  cores = 1,
  parallel = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{dem_files}{Um vetor de strings com os caminhos dos arquivos DEM a serem processados.}

\item{outdir}{O diretório onde os resultados serão salvos.}

\item{prefixo}{Um prefixo a ser adicionado aos nomes dos arquivos de saída.}

\item{cores}{Número de núcleos a serem usados (não utilizado nesta versão).}

\item{parallel}{Um booleano que indica se o processamento deve ser feito em paralelo (não utilizado nesta versão).}

\item{verbose}{Um booleano que indica se mensagens de progresso devem ser exibidas.}
}
\value{
Nenhum valor é retornado, mas os resultados são salvos no diretório especificado.
}
\description{
Esta função processa uma lista de arquivos DEM aplicando várias funções de análise
e salvando os resultados em um diretório especificado.
}
