% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reproj_raster}
\alias{reproj_raster}
\title{Reprojeta um raster de coordenadas geográficas para UTM.}
\usage{
reproj_raster(caminho_raster)
}
\arguments{
\item{caminho_raster}{Um caractere representando o caminho do arquivo raster a ser reprojetado.}
}
\value{
Um objeto raster reprojetado.
}
\description{
Esta função lê um arquivo raster, reprojeta-o de coordenadas geográficas
(EPSG:4326 ou EPSG:4674) para a zona UTM correspondente e retorna o raster reprojetado.
}
\examples{
# Reprojetar um raster específico
raster_reprojetado <- reproj_raster("caminho/para/raster.tif")

}
