% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{reprojetar_rasters}
\alias{reprojetar_rasters}
\title{Reprojeta todos os rasters em uma pasta de coordenadas geográficas para UTM.}
\usage{
reprojetar_rasters(path_entrada, path_saida)
}
\arguments{
\item{path_entrada}{Um caractere representando o caminho da pasta contendo os rasters
em coordenadas geográficas.}

\item{path_saida}{Um caractere representando o caminho da pasta onde os rasters
reprojetados serão salvos.}
}
\value{
Um vetor de caracteres com os nomes dos arquivos reprojetados.
}
\description{
Esta função lê todos os arquivos raster em uma pasta especificada, reprojeta
cada um deles de coordenadas geográficas (EPSG:4326 ou EPSG:4674) para a
zona UTM correspondente e salva os rasters reprojetados em uma nova pasta.
}
\examples{
# Reprojetar rasters de uma pasta para UTM
reprojetar_rasters("caminho/para/pasta/input", "caminho/para/pasta/output")

}
