% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{setup_saga}
\alias{setup_saga}
\title{Set Up the SAGA Environment}
\usage{
setup_saga(cores = 1, parallel = FALSE)
}
\arguments{
\item{cores}{Number of cores to use for processing. Default is 1.}

\item{parallel}{Logical indicating whether to enable parallel processing. Default is FALSE.}
}
\value{
A list representing the SAGA environment settings.
}
\description{
Initializes the SAGA GIS environment with specified settings for parallel processing.
}
