% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{write_saga_to_tiff}
\alias{write_saga_to_tiff}
\title{Convert SAGA Files to TIFF}
\usage{
write_saga_to_tiff(dir_path, outdir, prefix, num_tile)
}
\arguments{
\item{dir_path}{The directory path containing .sdat files.}

\item{outdir}{The output directory where .tif files will be saved.}

\item{prefix}{A prefix to add to the output file names.}
}
\value{
None. The function is used for its side effects of converting files.
}
\description{
Converts .sdat files in a specified directory to .tif format using the terra package.
}
