% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funcoes_rsaga.R
\name{process_dem}
\alias{process_dem}
\title{Main Processing Function with Parallel Execution}
\usage{
process_dem(dem, outdir, prefixo, cores = 1, parallel = FALSE, verbose = TRUE)
}
\arguments{
\item{dem}{The file path to the DEM.}

\item{outdir}{The output directory for the processed files.}

\item{cores}{The number of cores to use for parallel processing. Default is 2.}

\item{parallel}{Logical indicating whether to enable parallel processing. Default is TRUE.}

\item{verbose}{Logical indicating whether to print progress messages. Default is TRUE.}

\item{prefix}{string to be add at the begging of all file name}
}
\value{
None. The function is used for its side effects of processing the DEM.
}
\description{
Processes a DEM using multiple terrain analysis functions in parallel.
}
\examples{
# Define the path to the DEM file and the output directory
dem_path <- system.file("exdata", "mde_utm.tif", package = "yourPackageName")
output_directory <- tempdir()  # Use a temporary directory for output

# Process the DEM with default parameters
process_dem(dem = dem_path, outdir = output_directory, prefix = "morpho")

# Check the output files in the output directory
list.files(output_directory, pattern = "morpho", full.names = TRUE)
unlink(output_directory, recursive = TRUE)
}
